refactor: new folder for dcql
