+++ dif-presentation-exchange/Cargo.toml
@@ -17,4 +17,3 @@ serde.workspace = true
 serde_json.workspace = true
 serde_with.workspace = true
 url.workspace = true
-validator.workspace = true
