+++ dif-presentation-exchange/src/lib.rs
@@ -1,6 +1,4 @@
-pub mod claims;
 pub mod input_evaluation;
-pub mod meta;
 pub mod presentation_definition;
 pub mod presentation_submission;
