+++ oid4vp/Cargo.toml
@@ -31,3 +31,4 @@ serde_json.workspace = true
 serde_with.workspace = true
 tokio.workspace = true
 url.workspace = true
+validator.workspace = true
