+++ oid4vp/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod authorization_request;
+pub mod dcql;
 pub mod oid4vp;
 pub mod oid4vp_params;
 pub mod token;
